Style and UI gallery (#29)
* Quick cleanup

* Container and tweaks to notifications

* Include posts, reposts, buttons and inline results in devcards

* WIP: add progress bar

* Progress bars

* Added a go loop to decrease the testing timers

* Panel sizing was off

* Initial Verification Hub styles

* Fixed #31, running multiple node processes in para

in parallel.  This seems to require a node library in node world.  the
bash & and && don't cut it for us, as we want both to exit, too, when
ctrl+c or something similar is pressed.

* fixed #28: cider version

Cider and emacs were not playing nicely with repls started from `npm run
watch`, this other version of cider-nrepls integrates them nicely again.

* Fixed #25 - dev user

the url `http://localhost:8080/?dev=true` will now set up the state to
include a dev user.

Co-authored-by: niko <niko@niko.io>